fix: 파일 jsx로 수정
