--- src/pages/aboutus.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import Main from "../components/Aboutus/Aboutus";
-import Header from "../components/Header/Header";
-import Footer from "../components/Footer/Footer";
-
-
-function Aboutus() {
-    return (
-        <div>
-            <Header></Header>
-            <Main></Main>
-            <Footer></Footer>
-        </div>
-    )
-}
-
-export default Aboutus;
