+++ src/pages/applicationResult.jsx
@@ -1,13 +1,13 @@
-import Footer from "../components/Footer/Footer";
-import Header from "../components/Header/Header";
-import Main from "../components/Applies/ApplicationResult";
+import Footer from '../components/Footer/Footer';
+import Header from '../components/Header/Header';
+import Main from '../components/Applies/ApplicationResult';
 
 export default function ApplicationResult() {
-    return (
-        <div>
-            <Header />
-            <Main />
-            <Footer />
-        </div>
-    );
+  return (
+    <div>
+      <Header />
+      <Main />
+      <Footer />
+    </div>
+  );
 }
