--- src/pages/apply.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import Main from "../components/Apply/Apply";
-import Header from "../components/Header/Header";
-import Footer from "../components/Footer/Footer";
-
-function Apply() {
-    return (
-      <div>
-        <Header></Header>
-        <Main></Main>
-        <Footer></Footer>
-      </div>
-    );
-}
-
-export default Apply;
