--- src/pages/home.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import Main from "../components/Home/Home";
-import Header from "../components/Header/Header";
-import Footer from "../components/Footer/Footer";
-
-function Home() {
-    return (
-        <div>
-            <Header></Header>
-            <Main></Main>
-            <Footer></Footer>
-        </div>
-    )
-}
-
-export default Home;
