--- src/pages/members.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import Main from "../components/Members/Members";
-import Header from "../components/Header/Header";
-import Footer from "../components/Footer/Footer";
-
-function Members() {
-    return (
-        <div>
-            <Header></Header>
-            <Main></Main>
-            <Footer></Footer>
-        </div>
-    )
-}
-
-export default Members;
